% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot-threejs.r
\name{biplotEsaThreejs}
\alias{biplotEsaThreejs}
\title{Interactive 3D eigenstructure analysis biplot using three.js}
\usage{
biplotEsaThreejs(x, center = 4, g = 1, h = 1, ...)
}
\arguments{
\item{x}{A \code{repgrid} object.}

\item{center}{Centering of the grid data. \code{0} = no centering, \code{1} = row mean
(default), \code{2} = column mean, \code{3} = double centering, \code{4} = midpoint
centering.}

\item{g}{Power of the singular values assigned to the construct coordinates
(default \code{0}).}

\item{h}{Power of the singular values assigned to the element coordinates
(default \code{1 - g}).}

\item{...}{Additional arguments passed to \code{\link[=calcBiplotCoords]{calcBiplotCoords()}}.}
}
\value{
The file path of the generated HTML file (invisibly).
}
\description{
Wrapper around \code{\link[=biplotThreejs]{biplotThreejs()}} with default settings for eigenstructure
analysis biplot (\code{g = 1}, \code{h = 1}, \code{center = 4}).
}
\examples{
\dontrun{
biplotEsaThreejs(boeker)
}
}
\seealso{
\code{\link[=biplotThreejs]{biplotThreejs()}}, \code{\link[=biplotEsa3d]{biplotEsa3d()}}
}
