% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{cascade}
\alias{cascade}
\title{make ascending and descending vector}
\usage{
cascade(n, type = 1)
}
\arguments{
\item{n}{\code{integer} The length of the indexes}

\item{type}{(integer, default=1). If 1 the cascade index is returned.
2 returns the index of left and right side, 3 returns the length
of the left and right index vector}
}
\value{
vector (type 1 or 3) or list (type 2)
}
\description{
along a given length n make ascending indices until reaching
the midpoint and descend afterwards again.
}
\examples{
\dontrun{
for (n in 1:10) {
  print(cascade(n))
}
}
}
\keyword{internal}
