\name{clearRatings}
\alias{clearRatings}
\title{clearRatings...}
\usage{clearRatings(x, rows=NA, cols=NA, layer=1)
}
\description{clearRatings}
\details{set certain ratings in grid to NA (unknown)}
\value{\code{repgrid} object}
\keyword{internal}
\author{Mark Heckmann}
\arguments{\item{x}{repgrid object}
\item{rows}{rows to be set NA}
\item{cols}{columns to be set NA}
\item{layer}{layer of ratings to be set NA. Usually not important for the user (default = 1).}
}
\examples{\dontrun{

####  TODO  ####
}}

