% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgrid-output.r
\name{colorize_matrix_rows}
\alias{colorize_matrix_rows}
\title{Colorize matrix cell rows using crayon colors}
\usage{
colorize_matrix_rows(m, colors = "white", na.val = "white")
}
\arguments{
\item{m}{A matrix.}

\item{colors}{crayon colors as a string. One of
black, red, green, yellow, blue, magenta, cyan, white,
silver.}
}
\description{
Atomic matrices can be wrapped into crayon color codes without
destroying the structure or alignment. Used to indicate
preferred poles.
}
\examples{
m <- as.matrix(mtcars)
colorize_matrix_rows(m, "red")

}
\keyword{internal}
