% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.r
\name{constructPcaLoadings}
\alias{constructPcaLoadings}
\title{Extract loadings from PCA of constructs.}
\usage{
constructPcaLoadings(x)
}
\arguments{
\item{x}{\code{repgrid} object. This object is returned by the function \code{\link[=constructPca]{constructPca()}}.}
}
\value{
A matrix containing the factor loadings.
}
\description{
Extract loadings from PCA of constructs.
}
\examples{

p <- constructPca(bell2010)
l <- constructPcaLoadings(p)
l[1, ]
l[, 1]
l[1, 1]

}
