% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convertScivescoImportObjectToRepGridObject}
\alias{convertScivescoImportObjectToRepGridObject}
\title{Convert the returned object from the sci:vesco import function into a \code{repgrid}
object.}
\usage{
convertScivescoImportObjectToRepGridObject(import)
}
\arguments{
\item{x}{object returned from an import function.}
}
\value{
\code{repgrid} object.
}
\description{
Convert the returned object from the sci:vesco import function into a \code{repgrid}
object.
}
\author{
Mark Heckmann
}
\keyword{internal}

