% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-openrepgrid.r
\docType{data}
\name{data-bell2010}
\alias{data-bell2010}
\alias{bell2010}
\title{Grid data from Bell (2010).}
\description{
Grid data originated (but is not shown in the paper) from a study by Haritos, Gindinis, Doan and Bell (2004) on
element role titles. It was used to demonstrate the effects of construct alignment in Bell (2010, p. 46).
}
\examples{
bell2010
}
\references{
Bell, R. C. (2010).	A note on aligning constructs. \emph{Personal Construct Theory and Practice}, 7, 43-48.

Haritos, A., Gindidis, A., Doan, C., & Bell, R. C. (2004). The effect of element role titles on construct structure
and content. \emph{Journal of constructivist psychology, 17}(3), 221-236.
}
\keyword{data}
