\docType{data}
\name{data-bellmcgorry1992}
\alias{bellmcgorry1992}
\alias{data-bellmcgorry1992}
\title{Grid data from Bell and McGorry (1992).}
\description{
  The grid data set is used in Bell's technical report
  "Using SPSS to Analyse Repertory Grid Data" (1997, p. 6).
  Originally, the data comes from a study by Bell and
  McGorry (1992).
}
\references{
  Bell, R. C. (1977). \emph{Using SPSS to Analyse Repertory
  Grid Data}. Technical Report, University of Melbourne.

  Bell, R. C., & McGorry, P. (1992). The analysis of
  repertory grids used to monitor the perceptions of
  recovering psychotic patients. In A. Thomson & P. Cummins
  (Eds.), \emph{European Perspectives in Personal Construct
  Psychology} (p. 137-150).  Lincoln, UK: European Personal
  Construct Association.
}
\keyword{data}
\keyword{datasets}

