\docType{data}
\name{data-boeker}
\alias{boeker}
\alias{data-boeker}
\title{Grid data from Boeker (1996).}
\description{
  Grid data from a schizophrenic patient undergoing
  psychoanalytically oriented psychotherapy. The data was
  taken during the last stage of therapy (Boeker, 1996, p.
  163).
}
\references{
  Boeker, H. (1996). The reconstruction of the self in the
  psychotherapy of chronic schizophrenia: a case study with
  the Repertory Grid Technique. In: Scheer, J. W., Catina,
  A. (Eds.): \emph{Empirical Constructivism in Europe - The
  Personal Construct Approach} (p. 160-167).  Giessen:
  Psychosozial-Verlag.
}
\keyword{data}
\keyword{datasets}

