\name{data-fbb2003}
\alias{data-fbb2003}
\alias{data-fbb2003}
\title{Grid data from Fransella, Bell and Bannister (2003).}
\description{Grid data from Fransella, Bell and Bannister (2003).
A dataset used throughout the book "A Manual for Repertory Grid Technique" 
(Fransella, Bell and Bannister, 2003, p. 60).}
\alias{fbb2003}
\docType{data}
\references{Fransella, F., Bell, R. & Bannister, D. (2003). A Manual for Repertory 
Grid Technique (2. Ed.). Chichester: John Wiley & Sons.}
\keyword{data}

