\docType{data}
\name{data-feixas2004}
\alias{data-feixas2004}
\alias{feixas2004}
\title{Grid data from Feixas and Saul (2004).}
\description{
  A desription by the authors: "When Teresa, 22 years old,
  was seen by the second author (LAS) at the psychological
  services of the University of Salamanca, she was in the
  final year of her studies in chemical sciences. Although
  Teresa proves to be an excellent student, she reveals
  serious doubts about her self worth. She cries
  frequently, and has great difficulty in meeting others,
  even though she has a boyfriend who is extremely
  supportive. Teresa is anxiously hesitant about accepting
  a new job which would involve moving to another city 600
  Km away from home." (Feixas & Saul, 2004, p. 77).
}
\references{
  Feixas, G., & Saul, L. A. (2004). The Multi-Center
  Dilemma Project: an investigation on the role of
  cognitive conflicts in health.  \emph{The Spanish Journal
  of Psychology, 7}(1), 69-78.
}
\keyword{data}
\keyword{datasets}

