\docType{data}
\name{data-raeithel}
\alias{data-raeithel}
\alias{raeithel}
\title{Grid data from Raeithel (1998).}
\description{
  Grid data to demonstrate the use of Bertin diagrams
  (Raeithel, 1998, p. 223). The context of its
  administration is unknown.
}
\references{
  Raeithel, A. (1998). Kooperative Modellproduktion von
  Professionellen und Klienten. Erlaeutert am Beispiel des
  Repertory Grid.  In A. Raeithel (1998).
  Selbstorganisation, Kooperation, Zeichenprozess. Arbeiten
  zu einer kulturwissenschaftlichen, anwendungsbezogenen
  Psychologie (p. 209-254). Opladen: Westdeutscher Verlag.
}
\keyword{data}
\keyword{datasets}

