% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{data-slater1977a}
\alias{data-slater1977a}
\alias{slater1977a}
\title{Drug addict's grid data set from Slater (1977, p. 32).}
\description{
Drug addict's grid data set from Slater (1977, p. 32).
}
\references{
Slater, P. (1977). \emph{The measurement of intrapersonal space
             by grid technique}. London: Wiley.
}
\keyword{data}

