% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-openrepgrid.r
\docType{data}
\name{data-slater1977b}
\alias{data-slater1977b}
\alias{slater1977b}
\title{Grid data from Slater (1977).}
\description{
Grid data (ranked) from a seventeen year old female psychiatric patient (Slater, 1977, p. 110). She was depressed,
anxious and took to cutting herself. The data was originally reported by Watson (1970).
}
\examples{
slater1977b

}
\references{
Slater, P. (1977). \emph{The measurement of intrapersonal space by grid technique}. London: Wiley.

Watson, J. P. (1970). The relationship between a self-mutilating patient and her doctor. \emph{Psychotherapy and
Psychosomatics, 18}(1), 67-73.
}
\keyword{data}
