% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgridlist.R
\name{dim.repgridlist}
\alias{dim.repgridlist}
\title{Number of constructs and elements all repgrids in repgridlist}
\usage{
\method{dim}{repgridlist}(x)
}
\arguments{
\item{x}{A repgridist object.}
}
\value{
A list of length 2.
}
\description{
Number of constructs and elements all repgrids in repgridlist
}
\keyword{I}
\keyword{The}
\keyword{\code{ncol}}
\keyword{\code{nrow}}
\keyword{access}
\keyword{also}
\keyword{and}
\keyword{approach}
\keyword{are}
\keyword{as}
\keyword{avoid.}
\keyword{by}
\keyword{covered}
\keyword{dim}
\keyword{generic}
\keyword{internal}
\keyword{introduces}
\keyword{makiong}
\keyword{method.}
\keyword{not}
\keyword{others}
\keyword{so}
\keyword{the}
\keyword{them}
\keyword{this}
\keyword{to}
\keyword{try}
\keyword{warnings,}
\keyword{which}
