% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{dist_minmax}
\alias{dist_minmax}
\title{Calculate minimal and maximal possible distance}
\usage{
dist_minmax(x, along = 1, dmethod = "euclidean", p = 2, max.only = FALSE)
}
\description{
While the minimal distance will usually be zero, the maximal distance can be used to normalize arbitrary distances.
}
\keyword{internal}
