% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgrid-basicops.r
\name{doubleEntry}
\alias{doubleEntry}
\title{Join the constructs of a grid with the same reversed constructs.}
\usage{
doubleEntry(x)
}
\arguments{
\item{x}{\code{repgrid} object}
}
\value{
\code{repgrid} object
}
\description{
Join the constructs of a grid with the same reversed constructs.
}
\examples{
\dontrun{

     data(bell2010)
     doubleEntry(bell2010)
}

}
\keyword{internal}
