% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgrid-elements.r
\name{elementNames}
\alias{elementNames}
\alias{elementNames<-}
\title{Get or replace element names}
\usage{
elementNames(x)

elementNames(x, position) <- value
}
\arguments{
\item{x}{A repgrid object.}

\item{value}{Character vector of element names.}
}
\description{
Allows to get and set element names. 
Replaces the older functions \code{getElementNames}, \code{getElementNames2},
and \code{eNames} which are deprecated.
}
\examples{

## get element names
e <- elementNames(x)
e

## replace element names
elementNames(x) <- rev(e)       # reverse all element names
elementNames(x)[1] <- "Hannes"  # replace name of first element

# replace names of elements 1 and 3
elementNames(x)[c(1,3)] <- c("element 1", "element 3")   
 
}
