\name{getNoOfConstructs}
\alias{getNoOfConstructs}
\title{get number of constructs}
\usage{
  getNoOfConstructs(x)
}
\arguments{
  \item{x}{\code{repgrid} object}
}
\value{
  \code{numeric}
}
\description{
  get number of constructs
}
\examples{
\dontrun{

     getNoOfConstructs(bell2010)
}
}
\author{
  Mark Heckmann
}
\keyword{internal}

