\name{getNoOfElements}
\alias{getNoOfElements}
\title{get number of elements}
\usage{
  getNoOfElements(x)
}
\arguments{
  \item{x}{\code{repgrid} object}
}
\value{
  \code{numeric}
}
\description{
  get number of elements
}
\examples{
\dontrun{

     getNoOfElements(bell2010)
}
}
\author{
  Mark Heckmann
}
\keyword{internal}

