\name{getScale}
\alias{getScale}
\title{Get minimum and maximum scale value used in grid.}
\usage{getScale(x, output=1)
}
\description{Get minimum and maximum scale value used in grid. The values are 
returned either as a vector or a list.}
\value{Vector or list (depends on \code{output} containing 
minimum and maximum scale value.}
\keyword{internal}
\author{Mark Heckmann}
\arguments{\item{x}{\code{repgrid} object.}
\item{output}{Type of output object. 1= named vector, 2 = list.}
}

