\name{getScaleMidpoint}
\alias{getScaleMidpoint}
\title{get midpoint of scale...}
\usage{getScaleMidpoint(x)
}
\description{get midpoint of scale}
\value{\code{numeric}. Midpoint of scale.}
\keyword{internal}
\author{Mark Heckmann}
\arguments{\item{x}{\code{repgrid} object}
}
\examples{\dontrun{

getNoOfElements(bell2010)
}}

