% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.r
\name{grid_to_df_long}
\alias{grid_to_df_long}
\title{Export a grid to dataframe with long format}
\usage{
grid_to_df_long(x)
}
\arguments{
\item{x}{A \code{repgrid} object.}
}
\description{
Export a grid to dataframe with long format
}
\keyword{internal}
