% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.r
\name{grid_to_df_wide}
\alias{grid_to_df_wide}
\title{Export a grid to dataframe with wide format}
\usage{
grid_to_df_wide(x)
}
\arguments{
\item{x}{A \code{repgrid} object.}
}
\description{
Export a grid to dataframe with wide format
}
\keyword{internal}
