% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridlist.R
\name{gridlist}
\alias{gridlist}
\alias{is.gridlist}
\alias{as.gridlist}
\title{Add repgrids into a gridlist}
\usage{
gridlist(...)

is.gridlist(x)

as.gridlist(x)
}
\arguments{
\item{...}{Objects to be converted into `gridlist`}

\item{x}{Any object.}
}
\description{
Add repgrids into a gridlist

Test or create object of class `gridlist`
}
