% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.r
\name{indexBias}
\alias{indexBias}
\title{Calculate 'bias' of grid as defined by Slater (1977).}
\usage{
indexBias(x, min = NULL, max = NULL, digits = 2)
}
\arguments{
\item{x}{\code{repgrid} object.}

\item{min, max}{Minimum and maximum grid scale values. Nor needed if they are set for the grid.}

\item{digits}{Numeric. Number of digits to round to (default is \code{2}).}
}
\value{
Numeric.
}
\description{
"Bias records a tendency for responses to accumulate at one end of the grading scale" (Slater, 1977, p.88).
}
\note{
STATUS: Working and checked against example in Slater, 1977, p. 87.
}
\examples{
indexBias(boeker)

}
\references{
Slater, P. (1977). \emph{The measurement of intrapersonal space by Grid technique}. London: Wiley.
}
\seealso{
\code{\link[=indexVariability]{indexVariability()}}
}
