% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.r
\name{indexConflict2}
\alias{indexConflict2}
\title{Conflict measure for grids (Bassler et al., 1992) based on correlations.}
\usage{
indexConflict2(x, crit = 0.03)
}
\arguments{
\item{x}{A \code{repgrid} object.}

\item{crit}{Sensitivity criterion with which triads are marked as unbalanced. A bigger values will lead to less
imbalanced triads. The default is \code{0.03}. The value should be adjusted with regard to the researchers interest.}
}
\description{
The function calculates the conflict measure as devised by Bassler et al. (1992). It is an improved version of the
ideas by Slade and Sheehan (1979) that have been implemented in the function \code{\link[=indexConflict1]{indexConflict1()}}. The new approach
also takes into account the magnitude of the correlations in a trait to assess whether it is balanced or imbalanced.
As a result, small correlations that are psychologically meaningless are considered accordingly. Also, correlations
with a  small magnitude, i. e. near zero, which may  be positive or negative due to chance alone will no longer
distort the measure (Bassler et al., 1992).
}
\details{
Description of the balance / imbalance assessment:
\enumerate{
\item Order correlations of the triad by absolute magnitude, so that \eqn{ r_{max} > r_{mdn} > r_{min}}, \eqn{r_{max} > r_{mdn} > r_{min}}.
\item Apply Fisher's Z-transformation and division by 3 to yield values between 1 and -1  (\eqn{ Z_{max} > Z_{mdn} > Z_{min}, Z_{max} > Z_{mdn} > Z_{min}}).
\item Check whether the triad is balanced by assessing if the following relation holds:
\itemize{
\item If \eqn{Z_{max} Z_{mdn} > 0, Z_{max} x Z_{mdn} > 0},
the triad is balanced if \eqn{Z_{max} Z_{mdn} - Z_{min} <= crit},
\eqn{Z_{max} x Z_{mdn} - Z_{min} <= crit}.
\item If \eqn{Z_{max} Z_{mdn} < 0,  Z_{max} x Z_{mdn} < 0},
the triad is balanced if \eqn{Z_{min} - Z_{max} Z_{mdn} <= crit},
\eqn{Z_{min} - Z_{max} x Z_{mdn} <= crit}.
}
}
}
\section{Personal remarks (MH)}{
 I am a bit suspicious about step 2 from above. To devide by 3 appears pretty arbitrary.
The r for a z-values of 3 is 0.9950548 and not 1.
The r for 4 is 0.9993293. Hence, why not a value of 4, 5, or 6?
Denoting the value to devide by with \code{a}, the relation for the
first case translates into \eqn{a  Z_{max}  Z_{mdn} <= \frac{crit}{a} + Z_{min}},
\eqn{a x Z_{max} x Z_{mdn} =< crit/a + Z_{min}}. This shows that a bigger value of \code{a}
will make it more improbable that the relation will hold.
}

\examples{

indexConflict2(bell2010)

x <- indexConflict2(bell2010)
print(x)

# show conflictive triads
print(x, output = 2)

# accessing the calculations for further use
x$total
x$imbalanced
x$prop.balanced
x$prop.imbalanced
x$triads.imbalanced

}
\references{
Bassler, M., Krauthauser, H., & Hoffmann, S. O. (1992). A new approach to the identification of
cognitive conflicts in the repertory grid: An illustrative case study.
\emph{Journal of Constructivist Psychology, 5}(1), 95-111.

Slade, P. D., & Sheehan, M. J. (1979). The measurement of 'conflict' in repertory grids. \emph{British Journal of
Psychology, 70}(4), 519-524.
}
\seealso{
See \code{\link[=indexConflict1]{indexConflict1()}} for the older version of this measure; see \code{\link[=indexConflict3]{indexConflict3()}} for a measure based
on distances instead of correlations.
}
