% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.r
\name{indexDDI}
\alias{indexDDI}
\title{Dispersion of dependency index (DDI)}
\usage{
indexDDI(x, ds)
}
\arguments{
\item{x}{A \code{repgrid} object with \code{0}/\code{1} ratings only, where \code{1} indicates a dependency.}

\item{ds}{Predetermined size of sample of dependencies.}
}
\description{
A measure for the degree of dispersion of dependency in a situation-resource grid (dependency grid).
}
\details{
Applicable to dependency (situation-resource) grids (\code{0}/\code{1} ratings) only.

\emph{Caveat}: Value depends on chosen sample size \code{ds}, and the measurement range is not normalized (see Bell, 2001).
}
\examples{
# sample grid from Walker et al. (1988), p. 67
file <- system.file("extdata", "dep_grid_walker_1988_2.xlsx" , package = "OpenRepGrid")
x <- importExcel(file)

indexDDI(x, ds = 2:5) 

# using named vector
ds = c("2"=2, "3"=3, "4"=4, "5"=5)  
indexDDI(x, ds)
}
\references{
Bell, R. C. (2001). Some new Measures of the Dispersion of Dependency in a Situation-Resource Grid.
\emph{Journal of Constructivist Psychology, 14}(3), 227-234, \doi{doi:10.1080/713840106}.

Walker, B. M., Ramsey, F. L., & Bell, R. (1988). Dispersed and Undispersed Dependency.
\emph{International Journal of Personal Construct Psychology, 1}(1), 63-80, \doi{doi:10.1080/10720538808412765}.
}
\seealso{
\link{indexUncertainty}
}
