% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.r
\name{indexSelfRating}
\alias{indexSelfRating}
\title{Self Ratings
TBD}
\usage{
indexSelfRating(x, ideal, compare_to, max_diff = 2, midpoint_diff = 0)
}
\arguments{
\item{x}{A \code{repgrid} object.}

\item{ideal}{Ideal element to compare against (index or name).}

\item{compare_to}{Element to compare to ideal (index or name).}

\item{max_diff}{Max. difference from ideal to count as positive (numeric).}

\item{midpoint_diff}{TBD.}
}
\description{
Self Ratings
TBD
}
\examples{
indexSelfRating(boeker, "ideal self", "self")
indexSelfRating(boeker, "ideal self", "self", max_diff = 1)
}
