% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.r
\name{indexUncertainty}
\alias{indexUncertainty}
\title{Uncertainty index}
\usage{
indexUncertainty(x)
}
\arguments{
\item{x}{A \code{repgrid} object with \code{0}/\code{1} ratings only, where \code{1} indicates a dependency.}
}
\description{
A measure for the degree of dispersion of dependency in a situation-resource grid (dependency grid).
}
\details{
\emph{Note}: Applicable to dependency (situation-resource) grids (\code{0}/\code{1} ratings) only.
}
\examples{
# sample grid from Bell (2001, p.231)
file <- system.file("extdata", "dep_grid_bell_2001.xlsx" , package = "OpenRepGrid")
x <- importExcel(file)

indexUncertainty(x) 
}
\references{
Bell, R. C. (2001). Some new Measures of the Dispersion of Dependency in a Situation-Resource Grid.
\emph{Journal of Constructivist Psychology, 14}(3), 227-234, \doi{doi:10.1080/713840106}.
}
\seealso{
\link{indexDDI}
}
