\name{loadgrid}
\alias{loadgrid}
\title{load repertory grid file using a GUI...}
\description{load repertory grid file using a GUI}
\details{OpenRepGrid will try to guess what data type you are trying to load
by using the data format. Some formats are unique and are clearly 
attributable to a certain grid program (.scires, .grd) while others
are not.}
\keyword{internal}

