% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgrid.r
\name{makeEmptyRepgrid}
\alias{makeEmptyRepgrid}
\title{Constructor for repgrid class}
\usage{
makeEmptyRepgrid()
}
\value{
\code{repgrid} object
}
\description{
Constructor for repgrid class
}
\keyword{internal}
