% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgrid-basicops.r
\name{midpoint}
\alias{midpoint}
\title{Midpoint of the grid rating scale}
\usage{
midpoint(x)
}
\arguments{
\item{x}{\code{repgrid} object.}
}
\value{
Midpoint of scale.
}
\description{
Midpoint of the grid rating scale
}
\examples{
midpoint(bell2010)

}
