\name{modifyElement}
\alias{modifyElement}
\title{modifyElement...}
\usage{modifyElement(x, pos, name=NA, abbreviation=NA, status=NA, replace=FALSE)
}
\description{modifyElement}
\details{change the attributes of an element i.e. name, abbreviation, status etc.}
\value{\code{repgrid} object}
\keyword{internal}
\arguments{\item{x}{repgrid object}
\item{pos}{column number of element in the grid whose attributes are changed}
\item{name}{(optional) new name}
\item{abbreviation}{(optional) abbreviation of element name}
\item{status}{(optional) status of element (e.g. ideal etc.)}
\item{replace}{logical. wether to overwrite cuttent settings if NA provided}
}
\examples{\dontrun{

####  TODO  ####
}}

