% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.r
\name{[.openrepgridSettings}
\alias{[.openrepgridSettings}
\alias{print.openrepgridSettings}
\title{subset method for openrepgridSettings class}
\usage{
\method{[}{openrepgridSettings}(x, i, ...)

\method{print}{openrepgridSettings}(x, ...)
}
\description{
subset method for openrepgridSettings class

Print method for openrepgridSettings class
}
\keyword{internal}
