% Generated by roxygen2 (4.0.2): do not edit by hand
\name{orderBy}
\alias{orderBy}
\title{order one vector by another}
\usage{
orderBy(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{vector}
}
\value{
vector
}
\description{
small wrapper to order one vector by another, hardly worth a function
}
\keyword{internal}

