% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgrid-plots.r
\name{print.biplot2d}
\alias{print.biplot2d}
\title{Print method for biplot2d objects}
\usage{
\method{print}{biplot2d}(x, ...)
}
\arguments{
\item{x}{A \code{biplot2d} object returned by \code{\link[=biplot2d]{biplot2d()}}.}

\item{...}{Not used.}
}
\description{
Print method for biplot2d objects
}
