\name{print.constructD}
\alias{print.constructD}
\title{Print method for class constructD.}
\usage{
  \method{print}{constructD} (x, digits = 1,
    col.index = TRUE, ...)
}
\arguments{
  \item{x}{Object of class constructD.}

  \item{digits}{Numeric. Number of digits to round to
  (default is \code{2}).}

  \item{col.index}{Logical. Whether to add an extra index
  column so the column names are indexes instead of
  construct names. This option renders a neater output as
  long construct names will stretch the output (default is
  \code{TRUE}).}

  \item{...}{Not evaluated.}
}
\description{
  Print method for class constructD.
}
\keyword{internal}

