\name{print.constructPca}
\alias{print.constructPca}
\title{Print method for class constructPca.}
\usage{
  \method{print}{constructPca} (x, digits = 2, cutoff = 0,
    ...)
}
\arguments{
  \item{x}{Object of class constructPca.}

  \item{digits}{Numeric. Number of digits to round to
  (default is \code{2}).}

  \item{cutoff}{Loadings smaller than cutoff are not
  printed.}

  \item{...}{Not evaluated.}
}
\description{
  Print method for class constructPca.
}
\keyword{internal}

