\name{print.constructRmsCor}
\alias{print.constructRmsCor}
\title{Print method for class constructRmsCor.}
\usage{
  \method{print}{constructRmsCor} (x, digits = 2, ...)
}
\arguments{
  \item{x}{Object of class constructRmsCor.}

  \item{digits}{Numeric. Number of digits to round to
  (default is \code{2}).}

  \item{...}{Not evaluated.}
}
\description{
  Print method for class constructRmsCor.
}
\keyword{internal}

