% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.elementCor}
\alias{print.elementCor}
\title{Print method for class elementCor.}
\usage{
\method{print}{elementCor}(x, digits = 2, col.index = TRUE, upper = TRUE,
  ...)
}
\arguments{
\item{x}{Object of class elementCor}

\item{digits}{Numeric. Number of digits to round to (default is
\code{2}).}

\item{col.index}{Logical. Whether to add an extra index column so the
column names are indexes instead of construct names. This option
renders a neater output as long construct names will stretch
the output (default is \code{TRUE}).}

\item{upper}{Whether to display upper triangle of correlation matrix only
(default is \code{TRUE}).}

\item{...}{Not evaluated.}
}
\description{
Print method for class elementCor.
}
\keyword{internal}

