% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.r
\name{print.focus}
\alias{print.focus}
\title{Print method for focus objects}
\usage{
\method{print}{focus}(x, digits = 1, ...)
}
\arguments{
\item{x}{A \code{focus} object as returned by \code{\link[=focus]{focus()}}.}

\item{digits}{Number of decimal digits for matching scores (default \code{1}).}

\item{...}{Not used.}
}
\description{
Print method for focus objects
}
