% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.r
\name{print.indexBieri}
\alias{print.indexBieri}
\title{Print method for class indexBieri}
\usage{
\method{print}{indexBieri}(x, output = "I", digits = 3, ...)
}
\arguments{
\item{x}{Object of class \code{indexBieri}.}

\item{output}{String with each letter indicating which parts of the output
to print (default is \code{"IC"}, order does not matter): \code{I} = Information, \code{C}
= Matrix of matches.}

\item{digits}{Number of digits to display.}
}
\description{
Print method for class indexBieri
}
\keyword{internal}
