% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.r
\name{print.indexCognitiveProfile}
\alias{print.indexCognitiveProfile}
\title{Print method for indexCognitiveProfile}
\usage{
\method{print}{indexCognitiveProfile}(x, digits = 2, ...)
}
\description{
Print method for indexCognitiveProfile
}
\keyword{internal}
