% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.indexConflict1}
\alias{print.indexConflict1}
\title{Print function for class indexConflict1}
\usage{
\method{print}{indexConflict1}(x, digits = 1, ...)
}
\arguments{
\item{x}{Object of class indexConflict1.}

\item{digits}{Numeric. Number of digits to round to (default is
\code{1}).}

\item{...}{Not evaluated.}
}
\description{
Print function for class indexConflict1
}
\keyword{internal}

