\name{print.indexConflict3}
\alias{print.indexConflict3}
\title{print method for class indexConflict3}
\usage{
  \method{print}{indexConflict3} (x, digits = 2,
    output = 1, discrepancies = TRUE, ...)
}
\arguments{
  \item{x}{Output from funtion indexConflict3}

  \item{output}{Type of output. \code{output=1} will print
  all results to the console, \code{output=2} will only
  print the detailed statistics for elements and
  constructs.}

  \item{digits}{Numeric. Number of digits to round to
  (default is \code{2}).}

  \item{discrepancies}{Logical. Whether to show matrices of
  discrepancies in detailed element and construct stats
  (default \code{TRUE}).}

  \item{...}{Not evaluated.}
}
\description{
  print method for class indexConflict3
}
\keyword{internal}

