% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.r
\name{print.indexDilemma}
\alias{print.indexDilemma}
\title{Print method for class indexDilemma}
\usage{
\method{print}{indexDilemma}(x, digits = 2, ...)
}
\arguments{
\item{x}{Object of class indexDilemma}

\item{digits}{Numeric. Number of digits to round to (default is 
\code{2}).}

\item{...}{Not evaluated.}
}
\description{
Print method for class indexDilemma
}
\keyword{internal}
