% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.r
\name{print.indexDilemmatic}
\alias{print.indexDilemmatic}
\title{Print method for class indexDilemmatic}
\usage{
\method{print}{indexDilemmatic}(x, output = "SD", ...)
}
\arguments{
\item{x}{Object of class \code{indexDilemmatic}}

\item{output}{String with each letter indicating which parts of the output to print (default is \code{"SD"}, order
does not matter): \code{S} = Summary, \code{D} = Details (dilemmatic constructs).}
}
\description{
Print method for class indexDilemmatic
}
\keyword{internal}
