% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.r
\name{print.indexPolarization}
\alias{print.indexPolarization}
\title{Print method for class indexPolarization.}
\usage{
\method{print}{indexPolarization}(x, output = "ITCE", ...)
}
\arguments{
\item{x}{Object of class indexPolarization.}

\item{output}{String with each letter indicating which parts of the output to print (default is \code{"ITCE"}, order
does not matter): \code{I} = Information, \code{T} = Total Intensity, \code{C} = Constructs' intensities, \code{E} = Elements'
intensities.}
}
\description{
Print method for class indexPolarization.
}
\keyword{internal}
