% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.r
\name{print.indexSelfConstruction}
\alias{print.indexSelfConstruction}
\title{Print method for indexSelfConstruction}
\usage{
\method{print}{indexSelfConstruction}(x, digits = 2, ...)
}
\description{
Print method for indexSelfConstruction
}
\keyword{internal}
