% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.r
\name{print.org.matches}
\alias{print.org.matches}
\title{Print method for class org.matches.}
\usage{
\method{print}{org.matches}(
  x,
  output = "ICE",
  index = TRUE,
  names = TRUE,
  trim = 50,
  upper = TRUE,
  width = NA,
  ...
)
}
\arguments{
\item{x}{Object of class \code{org.matches}.}

\item{output}{String with each letter indicating which parts of the output to print (default is \code{"ICE"}, order
does not matter): \code{I} = Information, \code{C} = Constructs' matches, \code{E} = Elements' matches.}

\item{index}{Whether to add an index columns (default \code{TRUE}).}

\item{names}{names A vector of row names.}

\item{trim}{Trimmed length of names (default = \code{50}).}

\item{upper}{Whether to only show the upper triangle (default \code{TRUE}).}

\item{width}{Column width of output (numeric).}
}
\description{
Print method for class org.matches.
}
\keyword{internal}
