\name{print.statsElements}
\alias{print.statsElements}
\title{Print method for class statsElements}
\usage{
  \method{print}{statsElements} (x, digits = 2, ...)
}
\arguments{
  \item{x}{Object of class statsElements.}

  \item{digits}{Numeric. Number of digits to round to
  (default is \code{1}).}

  \item{...}{Not evaluated.}
}
\description{
  Print method for class statsElements
}
\keyword{internal}

