% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.r
\name{print_square_matrix}
\alias{print_square_matrix}
\title{Print a square matrix in well readable format}
\usage{
print_square_matrix(
  x,
  names = NA,
  trim = NA,
  index = TRUE,
  width = NA,
  upper = TRUE
)
}
\arguments{
\item{x}{A square \code{matrix}.}

\item{names}{names A vector of row names.}

\item{trim}{Trimmed length of names (numeric).}

\item{index}{Whether to add an index columns (default \code{TRUE}).}

\item{width}{Column width of output (numeric).}

\item{upper}{Whether to only show the upper triangle (default \code{TRUE}).}
}
\description{
Helper function to produce output for square matrices, e.g. distance
matrices etc. Adds a (trimmed) name and index column. Displays
upper triangle only by default. The function is used inside the
print method of several classes.
}
\keyword{internal}
