\name{randomWords}
\alias{randomWords}
\title{generate a random words...}
\usage{randomWords(n)
}
\description{generate a random words}
\details{randomWords generates a vector of random words taken from a small 
set of words}
\value{a string with n words (if length is not constrained)}
\keyword{internal}
\arguments{\item{n}{number of words to be generated (integer)}
}
\examples{randomWords(10)  # 10 random words}

