% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgrid-basicops.r
\name{ratings}
\alias{ratings}
\alias{ratings<-}
\title{Get ratings.}
\usage{
ratings(x, names = TRUE, trim = 10)

ratings(x, i, j) <- value
}
\arguments{
\item{x}{A \code{repgrid} object.}

\item{names}{Extract row and columns names (constructs and elements).}

\item{trim}{The number of characters a row or column name is trimmed to
(default is \code{10}). If \code{NA} no trimming is done. Trimming simply
saves space when displaying the output.}
}
\value{
A \code{matrix}.#'
}
\description{
Get ratings.
}
\examples{

## store Bell's dataset in x
x <- bell2010

## get ratings
ratings(x)


## replace ratings   
       
ratings(x)[1,1] <- 999
# noet that this is even simpler using the repgrid object directly
x[1,1] <- 888

#replace several values

ratings(x)[1,1:5] <- 999
x[1,1:5] <- 888   # the same

ratings(x)[1:3,5:6] <- matrix(55, 3, 2)
x[1:3,5:6] <- matrix(55, 3, 2)   # the same

}
\seealso{
\code{`[<--method`}
}
\author{
Mark Heckmann
}
