% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgrid-basicops.r
\name{reorder.repgrid}
\alias{reorder.repgrid}
\title{Invert construct and element order}
\usage{
\method{reorder}{repgrid}(x, what = "CE", ...)
}
\arguments{
\item{x}{A \code{repgrid} object.}

\item{what}{A string or numeric to indicate if constructs (\code{"C"}, \code{1}) or
elements (\code{"C"}, \code{1}), or both (\code{"CE"}, \code{12}) should be reversed.}

\item{...}{Ignored.}
}
\description{
Invert construct and element order
}
\examples{

# invert order of constructs
reorder(boeker, "C")
reorder(boeker, 1)

# invert order of elements
reorder(boeker, "E")
reorder(boeker, 2)

# invert both (default)
reorder(boeker)
reorder(boeker, "CE")
reorder(boeker, 12)

# not reordering
reorder(boeker, NA)

}
