% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridlist.R
\name{rep.repgrid}
\alias{rep.repgrid}
\title{Replicate repgrid objects}
\usage{
\method{rep}{repgrid}(x, n = 1, ...)
}
\arguments{
\item{x}{A `repgrid`` object.}

\item{n}{Number of times to replicate the grid.}
}
\value{
A `gridlist`` object.
}
\description{
Implements the \code{rep} method for \code{repgrid} objects.
}
\examples{
l <- rep(boeker, 3) # gridlist with 3 boeker grids
}
\keyword{internal}
