\name{repgrid-class}
\alias{repgrid-class}
\alias{repgrid}
\title{Definition of  repgrid class...}
\description{Definition of  repgrid class}
\author{Mark Heckmann}
\section{Slots}{\describe{\item{\code{meta}:}{}\item{\code{scale}:}{}\item{\code{elements}:}{}\item{\code{constructs}:}{}\item{\code{elicitation}:}{}\item{\code{ratings}:}{}\item{\code{coupled}:}{}\item{\code{calcs}:}{}\item{\code{plotdata}:}{}}}

