% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgridlist.R
\name{is.repgridlist}
\alias{is.repgridlist}
\alias{as.repgridlist}
\title{Test or create object of class \code{repgridlist}}
\usage{
is.repgridlist(x)

as.repgridlist(x)
}
\description{
Test or create object of class \code{repgridlist}
}
