\name{rglDrawConstructLabels}
\alias{rglDrawConstructLabels}
\title{draw constructs in rgl}
\usage{
  rglDrawConstructLabels(coords, labels = FALSE, dim = 1:3,
    c.cex = 0.6, c.text.col = grey(0.4), ...)
}
\arguments{
  \item{coords}{coordinates for constructs labels.}

  \item{labels}{labels for constructs.}

  \item{dim}{dimensions of coordinates to use.}

  \item{c.cex}{cex for construct text.}

  \item{c.text.col}{color for construct text.}

  \item{...}{not evaluated.}
}
\description{
  draw constructs in rgl
}
\author{
  Mark Heckmann
}
\keyword{internal}

