% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl-3d.r
\name{rglDrawConstructPoints}
\alias{rglDrawConstructPoints}
\title{draw constructs in rgl}
\usage{
rglDrawConstructPoints(coords, dim = 1:3, c.radius = 0.02,
  c.sphere.col = grey(0.4), ...)
}
\arguments{
\item{coords}{coordinates for construct points.}

\item{dim}{dimensions of coordinates to use.}

\item{c.radius}{radius of construct spheres.}

\item{c.sphere.col}{color of construct spheres.}

\item{...}{not evaluated.}
}
\description{
draw constructs in rgl
}
\author{
Mark Heckmann
}
\keyword{internal}
