% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl-3d.r
\name{rglDrawStandardEllipses}
\alias{rglDrawStandardEllipses}
\title{Draw standard ellipses in the origin in an rgl plot.}
\usage{
rglDrawStandardEllipses(max.dim = 1, lwd = 1, col = "black")
}
\arguments{
\item{max.dim}{soon}

\item{lwd}{soon}

\item{col}{soon}
}
\description{
Draw standard ellipses in the origin in an rgl plot.
}
\keyword{internal}
