% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.r
\name{saveAsWorksheet}
\alias{saveAsWorksheet}
\title{Add grids as sheets to an openxlsx Workbook}
\usage{
saveAsWorksheet(x, wb, format = "wide", sheet = "grid")
}
\arguments{
\item{x}{A \code{repgrid} object or a list of grids.}

\item{wb}{A \href{https://CRAN.R-project.org/package=openxlsx}{openxlsx} \code{Workbook} object.}

\item{format}{Two output formats are supported: \code{wide} (default) where each column represents one element, each row
represent one constructs (a common grid representation), and \code{long} where each row contains an element-construct
combination and the corresponding rating value. See \code{\link[=importExcel]{importExcel()}} for details and examples.}

\item{sheet}{Sheet name (defaults to \code{grid}). If \code{x} is a list of grids, the name of the list entry is used.
If it has no name, a sequential index is appended to \code{sheet}.}
}
\value{
Invisibly returns Workbook object.
}
\description{
\code{saveAsWorksheet} will add one or more grids to an a \href{https://CRAN.R-project.org/package=openxlsx}{openxlsx}
\code{Workbook} object.
}
\examples{
library(openxlsx)

wb <- createWorkbook()

# add grid to worksheet
saveAsWorksheet(boeker, wb, sheet = "boeker")

# add several grids (2nd has a name)
l <- list(boeker, "feixas' grid" = feixas2004) #
saveAsWorksheet(l, wb) # name sheet

file <- tempfile(fileext = ".xlsx")
saveWorkbook(wb, file)

\dontrun{
  browseURL(file) # may not work on all systems
}
}
\seealso{
\code{\link[=saveAsExcel]{saveAsExcel()}}
}
