\name{setMeta}
\alias{setMeta}
\title{setMeta...}
\usage{setMeta(x, type, id, name)
}
\description{setMeta}
\details{set meta data of a grid (e.g. id, name of interview partner)}
\value{\code{repgrid} object}
\keyword{internal}
\author{Mark Heckmann}
\arguments{\item{x}{repgrid object}
\item{type}{typemof grid in use (rating, ranked, implication)}
\item{id}{id of the interview}
\item{name}{name of the interview partner}
}
\examples{\dontrun{

####  TODO  ####
}}

