% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.r
\name{settingsLoad}
\alias{settingsLoad}
\title{Load OpenRepGrid settings}
\usage{
settingsLoad(file)
}
\arguments{
\item{file}{Path of the file to be loaded.}
}
\description{
OpenRepGrid settings saved in an a settings file with
the extension \code{.orgset} can be loaded to restore the
settings.
}
