% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.r
\name{settingsSave}
\alias{settingsSave}
\title{Save OpenRepGrid settings}
\usage{
settingsSave(file)
}
\arguments{
\item{file}{Path of the file to be saved to.}
}
\description{
The current settings of OpenRepGrid can be saved into a file with
the extension \code{.orgset}.
}
