% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgrid-output.r
\docType{methods}
\name{show,repgrid-method}
\alias{show,repgrid-method}
\title{Show method for repgrid}
\usage{
\S4method{show}{repgrid}(object)
}
\arguments{
\item{object}{A \code{repgrid} object.}
}
\description{
Show method for repgrid
}
