\name{showMeta}
\alias{showMeta}
\title{showMeta...}
\usage{showMeta(x)
}
\description{showMeta}
\details{prints meta information about the grid to the console (id, name of interviewee etc.)}
\value{\code{NULL}}
\keyword{internal}
\author{Mark Heckmann}
\arguments{\item{x}{repgrid object}
}
\examples{\dontrun{

####  TODO  ####
}}

