\name{showScale}
\alias{showScale}
\title{Print scale range information to the console.}
\usage{showScale(x)
}
\description{Print scale range information to the console.}
\value{\code{NULL}.}
\keyword{internal}
\author{Mark Heckmann}
\arguments{\item{x}{\code{repgrid} object.}
}
\examples{\dontrun{

showScale(raeithel)
showScale(bell2010)
}}

