% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgrid-basicops.r
\name{stop_if_not_0_1_ratings_only}
\alias{stop_if_not_0_1_ratings_only}
\title{Raise error if repgrid object has ratings other than 0/1}
\usage{
stop_if_not_0_1_ratings_only(x, name = "x")
}
\arguments{
\item{x}{A \code{repgrid} object.}

\item{name}{Name of tested variable in error message.}
}
\description{
Raise error if repgrid object has ratings other than 0/1
}
\keyword{internal}
