% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgrid-basicops.r
\name{stop_if_not_in_element_range}
\alias{stop_if_not_in_element_range}
\title{Raise error if element index is outside of range or element name is unknown}
\usage{
stop_if_not_in_element_range(x, element)
}
\arguments{
\item{x}{A `repgrid``` object.}

\item{element}{Element index or name.}
}
\description{
Raise error if element index is outside of range or element name is unknown
}
\keyword{internal}
