% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgrid-basicops.r
\name{stop_if_not_is_repgrid}
\alias{stop_if_not_is_repgrid}
\title{Raise error if not is repgrid object}
\usage{
stop_if_not_is_repgrid(x, name = "x")
}
\arguments{
\item{x}{Object to test. Raises an error if no \code{repgrid} object.}

\item{name}{Name of tested variable in error message.}
}
\description{
Raise error if not is repgrid object
}
\keyword{internal}
