% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgrid-basicops.r
\name{stop_if_scale_not_defined}
\alias{stop_if_scale_not_defined}
\title{Raise error if min or max of rating scale are undefined}
\usage{
stop_if_scale_not_defined(x)
}
\arguments{
\item{x}{Object to test. Raises an error the mina and max of the ratings scale are not defined}
}
\description{
Raise error if min or max of rating scale are undefined
}
\keyword{internal}
