\name{strReverse}
\alias{strReverse}
\title{reverse a string...}
\usage{strReverse(x)
}
\description{reverse a string}
\details{reverses the strings of a vector, i.e. c("ABC", "abc")
becomes c("CBA", "cba")}
\value{vector  a string or vector of strings with reversed chars}
\references{\url{http://www.mail-archive.com/r-help@r-project.org/msg102759.html}}
\keyword{internal}
\arguments{\item{x}{a string or a vector of strings}
}
\examples{strReverse(c("ABC", "abc"))}

