% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bindConstructs}
\alias{bindConstructs}
\title{Concatenate the constructs of two or more grids.}
\usage{
bindConstructs(..., index = FALSE)
}
\arguments{
\item{...}{One or more repgrid objects or a list containing
\code{repgrid} object.}

\item{index}{TODO. Logical (default \code{TRUE}). Whether to add an index at the end
               of each construct name so it remains clear from which grid each
               construct came.}
}
\value{
\code{repgrid} object with concatenated constructs.
}
\description{
I.e. the constructs are combined to form one long grid.
The girds must have the same set of elements and an identical
scale range. The order of the elements may differ.
}
\details{
This function can be used in order to analyse multiple grids
as one 'big grid' (eg. Slater, 1977, chap. 11).
}
\examples{
a <- randomGrid()
 b <- randomGrid()
 b@elements <- rev(a@elements)   # reverse elements
 bindConstructs(a, b)
 bindConstructs(a, b, a)

 # using lists of repgrid objects
 bindConstructs(a, list(a, b))
}
\author{
Mark Heckmann
}
\references{
Slater, P. (1977). \emph{The measurement of intrapersonal space
             by grid technique}. London: Wiley.
}

