% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.r
\name{calcAngles}
\alias{calcAngles}
\title{Calculate angles for points in first two columns.}
\usage{
calcAngles(x, dim = c(1, 2), clockwise = TRUE)
}
\arguments{
\item{x}{A matrix.}

\item{dim}{Dimensions used for calculating angles.}

\item{clockwise}{Logical. Positive angles are clockwise with x axis as basis.}
}
\value{
vector. The angles of each row point with the origin as reference.
}
\description{
The angles of the points given by the values in the first and second column of a matrix seen from the origin are
calculated in degrees.
}
\examples{
\dontrun{
m <- matrix(rnorm(9), 3)
calcAngles()
}
}
\keyword{internal}
