% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgrid-basicops.r
\name{clearRatings}
\alias{clearRatings}
\title{clear ratings}
\usage{
clearRatings(x, rows = NA, cols = NA, layer = 1)
}
\arguments{
\item{x}{repgrid object}

\item{rows}{rows to be set NA}

\item{cols}{columns to be set NA}

\item{layer}{layer of ratings to be set NA. Usually not important for the user (default = 1).}
}
\value{
\code{repgrid} object
}
\description{
set certain ratings in grid to NA (unknown)
}
\examples{
\dontrun{

   ####  TODO  ####
}
}
\keyword{internal}
