% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.r
\name{convertImportObjectToRepGridObject}
\alias{convertImportObjectToRepGridObject}
\title{convertImportObjectToRepGridObject.}
\usage{
convertImportObjectToRepGridObject(import)
}
\arguments{
\item{x}{object returned from an import function.}
}
\value{
\code{repgrid} object.
}
\description{
Convert the returned object from an import function into a \code{repgrid} 
object. Works for all importXInternal functions (except scivesco).
}
\keyword{internal}
