\name{convertImportObjectToRepGridObject}
\alias{convertImportObjectToRepGridObject}
\title{convertImportObjectToRepGridObject.}
\usage{convertImportObjectToRepGridObject(import)
}
\description{convertImportObjectToRepGridObject.
Convert the returned object from an import function into a \code{repgrid} 
object. Works for all importXInternal functions (except scivesco).}
\value{\code{repgrid} object.}
\keyword{internal}
\arguments{\item{x}{object returned from an import function.}
}

