% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.r
\name{convertScivescoImportObjectToRepGridObject}
\alias{convertScivescoImportObjectToRepGridObject}
\title{Convert the returned object from the sci:vesco import function into a \code{repgrid}
object.}
\usage{
convertScivescoImportObjectToRepGridObject(import)
}
\arguments{
\item{import}{List with relevant entries.}
}
\description{
Convert the returned object from the sci:vesco import function into a \code{repgrid}
object.
}
\keyword{internal}
