\name{convertScivescoImportObjectToRepGridObject}
\alias{convertScivescoImportObjectToRepGridObject}
\title{Convert the returned object from the sci:vesco import function into a repgrid...}
\usage{convertScivescoImportObjectToRepGridObject(import)
}
\description{Convert the returned object from the sci:vesco import function into a \code{repgrid} 
object.}
\value{\code{repgrid} object.}
\keyword{internal}
\author{Mark Heckmann}
\arguments{\item{x}{object returned from an import function.}
}

