% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-openrepgrid.r
\docType{data}
\name{data-fbb2003}
\alias{data-fbb2003}
\alias{fbb2003}
\title{Grid data from Fransella, Bell and Bannister (2003).}
\description{
A dataset used throughout the book "A Manual for Repertory Grid Technique" 
(Fransella, Bell and Bannister, 2003, p. 60).
}
\references{
Fransella, F., Bell, R. & Bannister, D. (2003). A Manual for Repertory 
             Grid Technique (2. Ed.). Chichester: John Wiley & Sons.
}
\keyword{data}
