\docType{data}
\name{data-leach2001}
\alias{data-leach2001}
\alias{leach2001a}
\alias{leach2001b}
\title{Pre- and post therapy dataset from Leach et al. (2001).}
\description{
  Case as described by the authors: "Sarah, aged 32, was
  referred with problems of depression and sexual
  difficulties relating to childhood sexual abuse. She had
  three children and was living with her male partner. From
  the age of 9, her brother, an adult, had sexually abused
  Sarah. She attended a group for survivors of child sexual
  abuse and completed repertory grids prior to the group,
  immediately after the group and at 3- and 6-month
  follow-up." (Leach et al. 2001, p. 230).\cr \cr
}
\details{
  \code{leach2001a} is the pre-therapy, \code{leach2001b}
  is the post-therapy therapy datset. The construct and
  elements are identical.
}
\references{
  Leach, C., Freshwater, K., Aldridge, J., & Sunderland, J.
  (2001). Analysis of repertory grids in clinical practice.
  \emph{The British Journal of Clinical Psychology, 40},
  225-248.
}
\keyword{data}
\keyword{datasets}

