% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-openrepgrid.r
\docType{data}
\name{data-mackay1992}
\alias{data-mackay1992}
\alias{mackay1992}
\title{Grid data from Mackay (1992).}
\description{
Data set 'Grid C' used in Mackay's paper on inter-element correlation (1992, p. 65).
}
\examples{
mackay1992

}
\references{
Mackay, N. (1992). Identification, reflection, and correlation: Problems in the bases of repertory grid
measures. \emph{International Journal of Personal Construct Psychology, 5}(1), 57-75.
}
\keyword{data}
