\name{data-slater1977a}
\alias{data-slater1977a}
\alias{data-slater1977a}
\title{Drug addict's grid data set from Slater (1977, p.}
\description{Drug addict's grid data set from Slater (1977, p. 32).}
\alias{slater1977a}
\docType{data}
\references{Slater, P. (1977). \emph{The measurement of intrapersonal space 
by grid technique}. London: Wiley.}
\keyword{data}

