% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridlist.R
\name{dim.gridlist}
\alias{dim.gridlist}
\title{Number of constructs and elements all repgrids in gridlist}
\usage{
\method{dim}{gridlist}(x)
}
\arguments{
\item{x}{A gridlist object.}
}
\value{
A list of length 2.
}
\description{
Number of constructs and elements all repgrids in gridlist
}
\details{
\code{nrow} and \code{ncol} are also covered by this approach
as they access the dim method. The others are not generic
and making them so introduces warnings, which I try to avoid.
}
\keyword{internal}
