% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgrid-basicops.r
\name{dim.repgrid}
\alias{dim.repgrid}
\title{Return size of a grid.}
\usage{
\method{dim}{repgrid}(x)
}
\arguments{
\item{x}{\code{repgrid} object.}
}
\value{
Numeric vector of length two with the number of 
             constructs and elements.
}
\description{
\code{dim} returns a numeric vector of length
two containing the number of constructs and elements.
}
\examples{
\dontrun{

     dim(bell2010)

}

}
\seealso{
\code{\link{getNoOfConstructs}};   \code{\link{getNoOfElements}}
}
\keyword{internal}
