\name{dim.repgrid}
\alias{dim.repgrid}
\title{Return size of a grid.}
\usage{dim.repgrid(x)
}
\description{Return size of a grid. \code{dim} returns a numeric vector of length
two containing the number of constructs and elements.}
\value{Numeric vector of length two with the number of 
constructs and elements.}
\keyword{internal}
\author{Mark Heckmann}
\seealso{\code{\link{getNoOfConstructs}};   \code{\link{getNoOfElements}}}
\arguments{\item{x}{\code{repgrid} object.}
}
\examples{\dontrun{

dim(bell2010)

}}

