\name{doubleEntry}
\alias{doubleEntry}
\title{Join the constructs of a grid with the same reversed constructs.}
\usage{
  doubleEntry(x)
}
\arguments{
  \item{x}{\code{repgrid} object}
}
\value{
  \code{repgrid} object
}
\description{
  Join the constructs of a grid with the same reversed
  constructs.
}
\examples{
\dontrun{

     data(bell2010)
     doubleEntry(bell2010)
}
}
\author{
  Mark Heckmann
}
\keyword{internal}

