\name{doubleEntry}
\alias{doubleEntry}
\title{Join the constructs of a grid with the same reversed constructs.}
\usage{doubleEntry(x)
}
\description{Join the constructs of a grid with the same reversed constructs.}
\value{\code{repgrid} object}
\keyword{internal}
\author{Mark Heckmann}
\arguments{\item{x}{\code{repgrid} object}
}
\examples{\dontrun{

data(bell2010)
doubleEntry(bell2010)
}}

