% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgrid-basicops.r
\name{[,repgrid-method}
\alias{[,repgrid-method}
\title{Extract parts of the repgrid object.}
\usage{
\S4method{[}{repgrid}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{A \code{repgrid} object.}

\item{i}{Row index (numeric)}

\item{j}{Column index. Either numeric or character (elements names).}

\item{...}{Not evaluated.}

\item{drop}{Not used.}
}
\description{
Methods for \code{"["}, i.e., subsetting of repgrid objects.
}
\examples{
x <- boeker
x[1:4, ]
x[, 1:3]
x[1:4, 1:3]
x[1:4, c("self", "ideal self", "mother")]

}
