\docType{methods}
\name{[}
\alias{[}
\alias{[,repgrid-method}
\title{Extract parts of the repgrid object.}
\description{
  Methods for \code{"["}, i.e., subsetting of repgrid
  objects.
}
\examples{
\dontrun{
   x <- randomGrid()
   x[1:4, ]
   x[ , 1:3]
   x[1:4,1:3]
   x[1,1]
}
}
\author{
  Mark heckmann
}
\keyword{internal}

