% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgrid-basicops.r
\docType{methods}
\name{[,repgrid,ANY,ANY-method}
\alias{[,repgrid,ANY,ANY-method}
\alias{[,repgrid-method}
\title{Extract parts of the repgrid object.}
\usage{
\S4method{[}{repgrid,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{A \code{repgrid} object.}

\item{i, j}{Row and column indices.}

\item{...}{Not evaluated.}

\item{drop}{Not used.}
}
\description{
Methods for \code{"["}, i.e., subsetting of repgrid objects.
}
\examples{

   x <- randomGrid()
   x[1:4, ] 
   x[ , 1:3] 
   x[1:4,1:3] 
   x[1,1]

}
\author{
Mark heckmann
}
