\name{extract-methods}
\alias{[,-method}
\alias{[}
\title{Extract parts of the repgrid object.}
\description{Extract parts of the repgrid object.
Methods for \code{"["}, i.e., subsetting of repgrid objects.}
\alias{[,repgrid-method}
\docType{methods}
\author{Mark heckmann}
\examples{\dontrun{

x <- randomGrid()
x[1:4, ] 
x[ , 1:3] 
x[1:4,1:3] 
x[1,1]
}}

