% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgrid-constructs.r
\name{getConstructNames}
\alias{getConstructNames}
\title{Get construct names}
\usage{
getConstructNames(x)
}
\arguments{
\item{x}{\code{repgrid} object.}
}
\description{
Get construct names
}
\section{Deprecated functions}{
 \code{getConstructNames()},
  and \code{cNames()} have been deprecated.
  Instead use \code{constructs()}.
}

\keyword{internal}
