\name{getElementNames}
\alias{getElementNames}
\title{Retrieve element names of repgrid object.}
\usage{getElementNames(x)
}
\description{Retrieve element names of repgrid object. Funtion si for convenience, 
so new users do not have to deal with object slots
as they will typically not have knowledge about R object 
structures (S3, S4).}
\value{vector   Vector containing the names of the elements.}
\keyword{internal}
\author{Mark Heckmann}
\arguments{\item{x}{\code{repgrid} object.}
}

