% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgrid-basicops.r
\name{getNoOfConstructs}
\alias{getNoOfConstructs}
\title{Get number of constructs}
\usage{
getNoOfConstructs(x)
}
\arguments{
\item{x}{\code{repgrid} object}
}
\value{
\code{numeric}
}
\description{
Get number of constructs
}
\examples{
\dontrun{

     getNoOfConstructs(bell2010)
}

}
\keyword{internal}
