\name{getNoOfConstructs}
\alias{getNoOfConstructs}
\title{get number of constructs...}
\usage{getNoOfConstructs(x)
}
\description{get number of constructs}
\value{\code{numeric}}
\keyword{internal}
\author{Mark Heckmann}
\arguments{\item{x}{\code{repgrid} object}
}
\examples{\dontrun{

getNoOfConstructs(bell2010)
}}

