% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgrid-basicops.r
\name{getNoOfElements}
\alias{getNoOfElements}
\title{Get number of elements}
\usage{
getNoOfElements(x)
}
\arguments{
\item{x}{\code{repgrid} object}
}
\value{
\code{numeric}
}
\description{
Get number of elements
}
\examples{
\dontrun{

     getNoOfElements(bell2010)
}

}
\keyword{internal}
