% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getNoOfElements}
\alias{getNoOfElements}
\title{Get number of elements}
\usage{
getNoOfElements(x)
}
\arguments{
\item{x}{\code{repgrid} object}
}
\value{
\code{numeric}
}
\description{
Get number of elements
}
\examples{
\dontrun{

     getNoOfElements(bell2010)
}
}
\author{
Mark Heckmann
}
\keyword{internal}

