\name{getNoOfElements}
\alias{getNoOfElements}
\title{get number of elements...}
\usage{getNoOfElements(x)
}
\description{get number of elements}
\value{\code{numeric}}
\keyword{internal}
\author{Mark Heckmann}
\arguments{\item{x}{\code{repgrid} object}
}
\examples{\dontrun{

getNoOfElements(bell2010)
}}

