% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgrid-basicops.r
\name{getRatingLayer}
\alias{getRatingLayer}
\title{get rating layer}
\usage{
getRatingLayer(x, layer = 1, names = TRUE, trim = 10)
}
\arguments{
\item{x}{\code{repgrid} object.}

\item{layer}{layer to be returned.}

\item{names}{extract row and columns names (constructs and elements).}

\item{trim}{the number of characters a row or column name is trimmed to 
(default is \code{10}). If \code{NA} no trimming is done. Trimming
simply saves space when displaying the output.}
}
\value{
a \code{matrix}
}
\description{
get rating layer
}
\examples{
\dontrun{

     getRatingLayer(bell2010)
}

}
\author{
Mark Heckmann
}
\keyword{internal}
