% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgrid-basicops.r
\name{getScale}
\alias{getScale}
\title{Get minimum and maximum scale value used in grid.}
\usage{
getScale(x, output = 1)
}
\arguments{
\item{x}{\code{repgrid} object.}

\item{output}{Type of output object. 1= named vector, 2 = list.}
}
\value{
Vector or list (depends on \code{output} containing 
                 minimum and maximum scale value.
}
\description{
The values are returned either as a vector or a list.
}
\keyword{internal}
