% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgrid-basicops.r
\name{getScaleMidpoint}
\alias{getScaleMidpoint}
\title{Get midpoint of the grid rating scale}
\usage{
getScaleMidpoint(x)
}
\arguments{
\item{x}{\code{repgrid} object.}
}
\value{
Midpoint of scale.
}
\description{
Get midpoint of the grid rating scale
}
\examples{
\dontrun{

     getScaleMidpoint(bell2010)

}

}
\author{
Mark Heckmann
}
\keyword{internal}
