\name{getScaleMidpoint}
\alias{getScaleMidpoint}
\title{Get midpoint of the grid rating scale...}
\usage{getScaleMidpoint(x)
}
\description{Get midpoint of the grid rating scale}
\value{Midpoint of scale.}
\keyword{internal}
\author{Mark Heckmann}
\arguments{\item{x}{\code{repgrid} object.}
}
\examples{\dontrun{

getScaleMidpoint(bell2010)

}}

