% Generated by roxygen2 (4.0.2): do not edit by hand
\name{indexBias}
\alias{indexBias}
\title{Calculate 'bias' of grid as defined by Slater (1977).}
\usage{
indexBias(x, min, max, digits = 2)
}
\arguments{
\item{x}{\code{repgrid} object.}

\item{min}{Minimum grid scale value.}

\item{max}{Maximum grid scale value.}

\item{digits}{Numeric. Number of digits to round to (default is
\code{2}).}
}
\value{
Numeric.
}
\description{
"Bias records a tendency for reponses to accumulate at one end of the
grading scale" (Slater, 1977, p.88).
}
\note{
STATUS:    Working and checked against example in Slater, 1977, p. 87.
}
\author{
Mark Heckmann
}
\references{
Slater, P. (1977). \emph{The measurement of intrapersonal space
                 by Grid technique}. London: Wiley.
}
\seealso{
\code{\link{indexVariability}}
}

