% Generated by roxygen2 (4.0.2): do not edit by hand
\name{indexVariability}
\alias{indexVariability}
\title{Calculate 'variability' of a grid as defined by Slater (1977).}
\usage{
indexVariability(x, min, max, digits = 2)
}
\arguments{
\item{x}{\code{repgrid} object.}

\item{min}{Minimum grid scale value.}

\item{max}{Maximum grdi scale value.}

\item{digits}{Numeric. Number of digits to round to (default is
\code{2}).}
}
\value{
Numeric.
}
\description{
Variability records a tendency for the responses to gravitate
towards both end of the gradings scale. (Slater, 1977, p.88).
}
\note{
STATUS: working and checked against example in Slater, 1977 , p.88.
}
\author{
Mark Heckmann
}
\references{
Slater, P. (1977). \emph{The measurement of intrapersonal space
               by Grid technique}. London: Wiley.
}
\seealso{
\code{\link{indexBias}}
}

