% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lapply_pb}
\alias{lapply_pb}
\title{lapply with a progress bar}
\usage{
lapply_pb(X, FUN, ...)
}
\arguments{
\item{X}{see ?lapply for parameter explanation}

\item{FUN}{see ?lapply}

\item{...}{see ?lapply}
}
\value{
list       see ?lapply
}
\description{
Can be used like standard base:::lapply. The only thing
it does is create an additional progress bar.
}
\examples{
\dontrun{

   l <- sapply(1:20000, function(x) list(rnorm(1000)))
   lapply_pb(l, mean)

}
}
\author{
Mark Heckmann
}
\seealso{
\code{\link{lapply}}
}
\keyword{internal}

