\name{loadgrid}
\alias{loadgrid}
\title{Load repertory grid file using a GUI}
\usage{
  loadgrid()
}
\description{
  OpenRepGrid will try to guess what data type you are
  trying to load by using the data format. Some formats are
  unique and are clearly attributable to a certain grid
  program (.scires, .grd) while others are not.
}
\keyword{internal}

