% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{map}
\alias{map}
\title{map a value onto others}
\usage{
map(x, each)
}
\arguments{
\item{x}{vector}

\item{each}{number of cuts}
}
\value{
vector
}
\description{
map a value onto others
}
\examples{
\dontrun{
map(1:10, 3)

m <- matrix(1:12, ncol = 4)
for (i in 1:12) {
  print(m[ring(i, 3), map(i, 4)])
}
}

}
\keyword{internal}
