% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgrid-basicops.r
\name{modifyConstruct}
\alias{modifyConstruct}
\title{modify a construct}
\usage{
modifyConstruct(x, pos, l.name = NA, l.preferred = NA, l.emerged = NA,
  r.name = NA, r.preferred = NA, r.emerged = NA, replace = FALSE)
}
\arguments{
\item{x}{repgrid object}

\item{pos}{row number of construct in the grid to be changed}

\item{l.name}{(optional) name of the left pole (string)}

\item{l.preferred}{(optional) is the left one the preferred pole? (logical)}

\item{l.emerged}{(optional) is the left one the emergent pole? (logical)}

\item{r.name}{(optional) name of the right pole (string)}

\item{r.preferred}{(optional) is the right one the preferred pole? (logical)}

\item{r.emerged}{(optional) is the right one the emergent pole? (logical)}

\item{replace}{should the sttributes be replaced if NA is provided?}
}
\value{
\code{repgrid} object
}
\description{
change the attributes of a construct
}
\examples{
\dontrun{

   ####  TODO  ####
}

}
\keyword{internal}
