% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgrid-basicops.r
\name{modifyElement}
\alias{modifyElement}
\title{modifyElement}
\usage{
modifyElement(
  x,
  pos,
  name = NA,
  abbreviation = NA,
  status = NA,
  replace = FALSE
)
}
\arguments{
\item{x}{repgrid object}

\item{pos}{column number of element in the grid whose attributes are changed}

\item{name}{(optional) new name}

\item{abbreviation}{(optional) abbreviation of element name}

\item{status}{(optional) status of element (e.g. ideal etc.)}

\item{replace}{logical. whether to overwrite current settings if NA provided}
}
\value{
\code{repgrid} object
}
\description{
change the attributes of an element i.e. name, abbreviation, status etc.
}
\examples{
\dontrun{

####  TODO  ####
}
}
\keyword{internal}
