% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.r
\name{normalize}
\alias{normalize}
\title{Normalize rows or columns by its standard deviation.}
\usage{
normalize(x, normalize = 0, ...)
}
\arguments{
\item{x}{\code{matrix}}

\item{normalize}{A numeric value indicating along what direction (rows, columns)
to normalize by standard deviations. \verb{0 = none, 1= rows, 2 = columns}
(default is \code{0}).}

\item{...}{Not evaluated.}
}
\value{
Not yet defined TODO!
}
\description{
Normalize rows or columns by its standard deviation.
}
\examples{
x <- matrix(sample(1:5, 20, rep = TRUE), 4)
normalize(x, 1) # normalizing rows
normalize(x, 2) # normalizing columns

}
