% Generated by roxygen2 (4.0.2): do not edit by hand
\name{normalize}
\alias{normalize}
\title{Normalize rows or columns by its standard deviation.}
\usage{
normalize(x, normalize = 0, ...)
}
\arguments{
\item{x}{\code{matrix}}

\item{normalize}{A numeric value indicating along what direction (rows, columns)
to normalize by standard deviations. \code{0 = none, 1= rows, 2 = columns}
(default is \code{0}).}

\item{...}{Not evaluated.}
}
\value{
Not yet definde TODO!
}
\description{
Normalize rows or columns by its standard deviation.
}
\examples{
\dontrun{

 x <- matrix(sample(1:5, 20, rep=T), 4)
 normalize(x, 1)						      # normalizing rows
 normalize(x, 2)						      # normalizing columns
}
}
\author{
Mark Heckmann
}

