\docType{package}
\name{OpenRepGrid-overview}
\alias{OpenRepGrid-overview}
\alias{OpenRepGrid-overview-package}
\title{\pkg{OpenRepGrid}: Annotated overview of package functions.}
\description{
  This doumentation page contains an overview over the
  package functions ordered by topics. The best place to
  start learning OpenRepGrid will be the package website
  \url{http://www.openrepgrid.org} though.
}
\section{Functions by topic}{
  \bold{Functions for manipulating grids} \cr \cr

  \tabular{ll}{ \code{\link{left}} \tab Move construct(s)
  to the left \cr \code{\link{right}} \tab Move
  construct(s) to the right \cr \code{\link{up}} \tab Move
  construct(s) upwards \cr \code{\link{down}} \tab Move
  construct(s) downwards \cr }

  \bold{Loading and saving data} \cr \cr

  \bold{Analyzing constructs} \cr \cr

  \bold{Analyzing elements} \cr \cr

  \bold{Visual representation} \cr \cr

  \bold{Index measures} \cr \cr

  \bold{Package Data} \cr \cr

  Type the name of the dataset to the console amd press
  enter to see the grid \cr

  \emph{Single grids} \cr

  \tabular{ll}{ \code{\link{bell2010}} \tab Grid data from
  a study by Haritos et al. (2004) on role titles; used for
  demonstration of construct alignment in Bell (2010, p.
  46). \cr \code{\link{bellmcgorry1992}} \tab Grid from a
  psychotic patient used in Bell (1997, p. 6). Data
  originated from a study by Bell and McGorry (1992). \cr
  \code{\link{boeker}} \tab Grid from seventeen year old
  female schizophrenic patient undergoing last stage of
  psychoanalytically oriented psychotherapy (Boeker, 1996,
  p. 163). \cr \code{\link{fbb2003}} \tab Dataset used in
  \emph{A manual for Repertory Grid Technique} (Fransella,
  Bell, & Bannister, 2003b, p. 60). \cr
  \code{\link{feixas2004}} \tab Grid from a 22 year old
  Spanish girl suffering self-worth problems (Feixas &
  Saul, 2004, p. 77). \cr \code{\link{mackkay1992}} \tab
  Dataset \emph{Grid C} used in Mackay's paper on
  inter-element correlation (1992, p. 65). \cr
  \code{\link{leach2001a, leach2001b}} \tab Pre- (a) and
  post-therapy (b) dataset from sexual child abuse survivor
  (Leach, Freshwater, Aldridge, & Sunderland, 2001, p.
  227). \cr \code{\link{raeithel}} \tab Grid data to
  demonstrate the use of Bertin diagrams (Raeithel, 1998,
  p. 223). The context of its administration is unknown.
  \cr \code{\link{slater1997a}} \tab Drug addict grid
  dataset from (Slater, 1977, p. 32). \cr
  \code{\link{slater1977b}} \tab grid dataset (ranked) from
  a seventeen year old female psychiatric patient (Slater,
  1977, p. 110) showing depression, anxiety and
  self-mutilation.  The data was originally reported by
  Watson (1970). \cr }

  \emph{Multiple grids} \cr

  NOT YET AVAILABLE \cr
}
\keyword{package}

