\name{[.openrepgridSettings}
\alias{[.openrepgridSettings}
\alias{print.openrepgridSettings}
\title{subset method for openrepgridSettings class}
\usage{
  \method{[}{openrepgridSettings} (x, i, ...)

  \method{print}{openrepgridSettings} (x, ...)
}
\description{
  subset method for openrepgridSettings class

  Print method for openrepgridSettings class
}
\keyword{internal}

