% Generated by roxygen2 (4.0.2): do not edit by hand
\name{orderByString}
\alias{orderByString}
\title{find the order of a string vector so it will match the order of another}
\usage{
orderByString(x, y)
}
\arguments{
\item{x}{a vector of strings}

\item{y}{a vector of strings}
}
\value{
a vector of strings
}
\description{
find the order of a string vector so it will match the order of another
}
\examples{
\dontrun{
   a <- c("c", "a", "b")
   b <- c("b", "c", "a")
   index <- orderByString(a, b)    # to order b like a needs what indexes?
   index
   b[index]
}
}
\keyword{internal}

