% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferred_poles.R
\name{preferredPolesByIdeal}
\alias{preferredPolesByIdeal}
\title{Set preferred pole by ideal element}
\usage{
preferredPolesByIdeal(x, ideal, none_range = NULL, align = FALSE)
}
\arguments{
\item{x}{A \code{repgrid} object.}

\item{ideal}{Index or name of ideal element.}

\item{none_range}{Range of ratings that do not allow assining a preferred pole (\code{NULL} be default).}

\item{align}{Align preferred poles on same side (default \code{FALSE}). See \code{\link[=alignByPreferredPoles]{alignByPreferredPoles()}}.}
}
\description{
The preferred construct pole is inferred from the rating of the ideal element.
The preferred pole is the side of the ideal element. If the ideal is rated on the
scale midpoint (or within \code{none_range}), none of the poles is preferred.
}
