% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{print.distance}
\alias{print.distance}
\title{Print method for class distance.}
\usage{
\method{print}{distance}(
  x,
  digits = 2,
  col.index = TRUE,
  upper = TRUE,
  diag = FALSE,
  cutoffs = NA,
  ...
)
}
\arguments{
\item{x}{Object of class distance.}

\item{digits}{Numeric. Number of digits to round to (default is
\code{2}).}

\item{col.index}{Logical. Whether to add an extra index column so the
column names are indexes instead of construct names. This option
renders a neater output as long construct names will stretch
the output (default is \code{TRUE}).}

\item{upper}{Whether to display upper triangle of correlation matrix only
(default is \code{TRUE}).}

\item{diag}{Whether to show the matrix diagonal.}

\item{cutoffs}{Cutoff values. Values below or above this interval are not
printed. For Slater distances \code{c(.8, 1.2)} are common
values.}

\item{...}{Not evaluated.}
}
\description{
Print method for class distance.
}
\keyword{internal}
