% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.r
\name{print.indexDilemma}
\alias{print.indexDilemma}
\title{Print method for class indexDilemma}
\usage{
\method{print}{indexDilemma}(x, digits = 2, output = "SPCD", ...)
}
\arguments{
\item{x}{Object of class indexDilemma}

\item{digits}{Numeric. Number of digits to round to (default is \code{2}).}

\item{output}{String with each letter indicating which parts of the output to print 
(default is \code{"OCD"}, order does not matter):
\code{S} = Summary (Number of IDs, PID, etc.),
\code{P} = Analysis parameters,
\code{C} = Construct classification table, 
\code{D} = Implicative dilemmas table.}

\item{...}{Not evaluated.}
}
\description{
Print method for class indexDilemma
}
\keyword{internal}
