% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.indexIntensity}
\alias{print.indexIntensity}
\title{Print method for class indexIntensity.}
\usage{
\method{print}{indexIntensity}(x, digits = 2, ...)
}
\arguments{
\item{x}{Object of class indexIntensity.}

\item{digits}{Numeric. Number of digits to round to (default is
\code{2}).}

\item{...}{Not evaluated.}
}
\description{
Print method for class indexIntensity.
}
\keyword{internal}

