% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.r
\name{print.indexIntensity}
\alias{print.indexIntensity}
\title{Print method for class indexIntensity.}
\usage{
\method{print}{indexIntensity}(x, digits = 2, output = "TCE", ...)
}
\arguments{
\item{x}{Object of class indexIntensity.}

\item{digits}{Numeric. Number of digits to round to (default is \code{2}).}

\item{output}{String with each letter indicating which parts of the output to print (default is \code{"TCE"}, order
does not matter): \code{T} = Total Intensity, \code{C} = Constructs' intensities, \code{E} = Elements' intensities.}
}
\description{
Print method for class indexIntensity.
}
\keyword{internal}
