% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.r
\name{print.rmsCor}
\alias{print.rmsCor}
\title{Print method for class rmsCor (RMS correlation for constructs or elements)}
\usage{
\method{print}{rmsCor}(x, digits = 2, ...)
}
\arguments{
\item{x}{Object of class rmsCor.}

\item{digits}{Numeric. Number of digits to round to (default is
\code{2}).}

\item{...}{Not evaluated.}
}
\description{
Print method for class rmsCor (RMS correlation for constructs or elements)
}
\keyword{internal}
