% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.statsConstructs}
\alias{print.statsConstructs}
\title{Print method for class statsConstructs}
\usage{
\method{print}{statsConstructs}(x, digits = 2, ...)
}
\arguments{
\item{x}{Object of class statsConstructs.}

\item{digits}{Numeric. Number of digits to round to (default is
\code{1}).}

\item{...}{Not evaluated.}
}
\description{
Print method for class statsConstructs
}
\keyword{internal}

