% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{randomSentence}
\alias{randomSentence}
\title{generate a random sentence with n words}
\usage{
randomSentence(n, maxchar = Inf)
}
\arguments{
\item{n}{number of word in sentence}

\item{maxchar}{maximal number of characters per sentence. Note that whole 
words (not part of words) are excluded if the maximal number 
 is exceeded.}
}
\value{
a string with n words (if length is not constrained)
}
\description{
generate a random sentence with n words
}
\examples{
 
randomSentence(10)   # one random sentence with 10 words
}
\keyword{internal}
