% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{randomWords}
\alias{randomWords}
\title{generate a random words}
\usage{
randomWords(n)
}
\arguments{
\item{n}{number of words to be generated (integer)}
}
\value{
a string with n words (if length is not constrained)
}
\description{
randomWords generates a vector of random words taken from a small 
set of words
}
\examples{
randomWords(10)  # 10 random words
}
\keyword{internal}
