% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rglDrawStandardAxes}
\alias{rglDrawStandardAxes}
\title{Draw standard axes in the origin in an rgl plot.}
\usage{
rglDrawStandardAxes(max.dim = 1, lwd = 1, a.cex = 1.1, a.col = "black",
  a.radius = 0.05, labels = TRUE, spheres = FALSE, ...)
}
\arguments{
\item{max.dim}{maximum length of axis.}

\item{lwd}{line width.}

\item{a.cex}{cex for axis labels.}

\item{a.col}{axis color.}

\item{a.radius}{radius of spheres at the end points of the axes.}

\item{labels}{logical. wether to draw axis labels.}

\item{spheres}{logical. wether to draw axis spheres at the end points.}

\item{...}{not evaluated.}
}
\description{
Draw standard axes in the origin in an rgl plot.
}
\author{
Mark Heckmann
}
\keyword{internal}

