\name{setMeta}
\alias{setMeta}
\title{setMeta}
\usage{
  setMeta(x, type, id, name)
}
\arguments{
  \item{x}{repgrid object}

  \item{type}{typemof grid in use (rating, ranked,
  implication)}

  \item{id}{id of the interview}

  \item{name}{name of the interview partner}
}
\value{
  \code{repgrid} object
}
\description{
  set meta data of a grid (e.g. id, name of interview
  partner)
}
\examples{
\dontrun{

   ####  TODO  ####
}
}
\author{
  Mark Heckmann
}
\keyword{internal}

