\name{settingsLoad}
\alias{settingsLoad}
\title{Load OpenRepGrid settings}
\usage{
  settingsLoad(file)
}
\arguments{
  \item{file}{Path of the file to be loaded. If a path is
  not supplied an interactive file chooser dialog is
  opened.}
}
\description{
  OpenRepGrid settings saved in an a settings file with the
  extension \code{.orgset} can be loaded to restore the
  settings.
}

