% Generated by roxygen2 (4.0.2): do not edit by hand
\name{settingsSave}
\alias{settingsSave}
\title{Save OpenRepGrid settings}
\usage{
settingsSave(file)
}
\arguments{
\item{file}{Path of the file to be saved to. If a path is not supplied
an interactive file saver dialog is opened.}
}
\description{
The current settings of OpenRepGrid can be saved into a file with
the extension \code{.orgset}.
}

