\name{showScale}
\alias{showScale}
\title{Print scale range information to the console.}
\usage{
  showScale(x)
}
\arguments{
  \item{x}{\code{repgrid} object.}
}
\value{
  \code{NULL}.
}
\description{
  Print scale range information to the console.
}
\examples{
\dontrun{

   showScale(raeithel)
   showScale(bell2010)
}
}
\author{
  Mark Heckmann
}
\keyword{internal}

