% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repgrid-basicops.r
\name{showScale}
\alias{showScale}
\title{Print scale range information to the console.}
\usage{
showScale(x)
}
\arguments{
\item{x}{\code{repgrid} object.}
}
\value{
\code{NULL}.
}
\description{
Print scale range information to the console.
}
\examples{
\dontrun{

   showScale(raeithel)
   showScale(bell2010)
}

}
\keyword{internal}
