% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{[<-,repgrid-method}
\alias{[<-,repgrid-method}
\title{Method for "<-" assignment of the repgrid ratings.}
\usage{
\S4method{[}{repgrid}(x, i, j, ...) <- value
}
\arguments{
\item{x}{A \code{repgrid} object.}

\item{i,j}{Row and column indices.}

\item{...}{Not evaluated.}

\item{value}{Numeric replacement value(s).}
}
\description{
It should be possible to use it for ratings on all layers.
}
\examples{
\dontrun{
   x <- randomGrid()
   x[1,1] <- 2
   x[1, ] <- 4
   x[ ,2] <- 3
}
}
\author{
Mark Heckmann
}

