% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trim_val}
\alias{trim_val}
\title{trim vector to lower or upper value}
\usage{
trim_val(x, minmax = c(-Inf, Inf), na = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{minmax}{minimal and maximal possible value (default c(-Inf, Inf)
i.e. no trimming occurs)}

\item{na}{Use NAs for replacing values that are out of range}
}
\value{
vector   vector whose elemenets that are out of range are replaced
}
\description{
the range a value may take is resticted to by an upper and
lower boundary. If it excedd the boundary the value is replaced
by the boundary value or alternatively by NA
}
\examples{
trim_val(30)
trim_val(30, c(10,20))
}
\keyword{internal}

